Moved some classes to the "right" package. Tuned thread pool configs a bit; also introduced environment variables to control the pool parameters without having to make code mods.
