+++ worker/src/main/resources/application.properties
@@ -19,9 +19,16 @@ report.progress.db.frequency=100
 report.progress.log.frequency=10000
 
 ## ---------------------------------------------------------------------------------------  PCP THREAD-POOL CONFIG
-pcp.core.pool.size=5
-pcp.max.pool.size=10
-pcp.queue.capacity=100
+## These properties apply to "patientProcessorThreadPool".
+pcp.core.pool.size=${AB2D_PCP_POOL_CORE_SIZE:#{10}}
+pcp.max.pool.size=${AB2D_PCP_POOL_MAX_SIZE:#{10}}
+pcp.queue.capacity=${AB2D_PCP_QUEUE_CAPACITY:#{100}}
+
+## ----------------------------------------------------------------------------- MAIN JOB PROCESSOR THREAD-POOL CONFIG
+## These properties apply to "mainJobProcessingPool".
+job.core.pool.size=${AB2D_JOB_POOL_CORE_SIZE:#{5}}
+job.max.pool.size=${AB2D_JOB_POOL_MAX_SIZE:#{10}}
+job.queue.capacity=${AB2D_JOB_QUEUE_CAPACITY:#{0}}
 
 
 ## ---------------------------------------------------------------------------------------------------   STUCK JOB
